refactor(session): extract execution control into src/session/execution/
Stepping, continue, pause and the thread list move to `ExecutionController`.
The three step methods differed only in a DAP command, an operation name, a
log tag and a success message, so those become a `STEP_KINDS` table over one
shared preamble; every log literal and `ProxyNotRunningError` operation name
is unchanged. `_executeStepOperation` becomes the public `executeStep`, which
is how its tests reach it now.
